Create Quartiles
Objective 
In this challenge, we practice calculating quartiles. Check out the Tutorial tab for learning materials and an instructional video!

Task 
Given an array, X, of n  integers, calculate the respective first quartile (Q1), second quartile (Q2), and third quartile (Q3). It is guaranteed that  Q1, Q2 , and Q3  are integers.

Input Format
The first line contains an integer, n , denoting the number of elements in the array. 
The second line contains n  space-separated integers describing the array's elements.

Output Format
Print 3 lines of output in the following order:
1. The first line should be the value of Q1.
2. The second line should be the value of Q2 .
3. The third line should be the value of Q3.