Create Interquartile Range
Objective 
In this challenge, we practice calculating the interquartile range. We recommend you complete the Quartiles challenge before attempting this problem.

Task 
The interquartile range of an array is the difference between its first (Q1) and third (Q2) quartiles (i.e., Q3 - Q1).
Given an array, X , of n  integers and an array,  F, representing the respective frequencies of X's elements, construct a data set, S , where each xi occurs at frequency fi . Then calculate and print S's interquartile range, rounded to a scale of  1 decimal place (i.e., 12.3 format).

Tip: Be careful to not use integer division when averaging the middle two elements for a data set with an even number of elements, and be sure to not include the median in your upper and lower data sets.

Input Format
The first line contains an integer, n, denoting the number of elements in arrays X  and F. 
The second line contains n  space-separated integers describing the respective elements of array X. 
The third line contains n  space-separated integers describing the respective elements of array F.

Output Format
Print the interquartile range for the expanded data set on a new line. Round your answer to a scale of  decimal place (i.e., format).