Create Standard Deviation
Objective 
In this challenge, we practice calculating standard deviation. Check out the Tutorial tab for learning materials and an instructional video!

Task 
Given an array, X , of N  integers, calculate and print the standard deviation. Your answer should be in decimal form, rounded to a scale of  1 decimal place (i.e., 12.3 format). An error margin of +- 0.1 will be tolerated for the standard deviation.

Input Format
The first line contains an integer, N, denoting the number of elements in the array. 
The second line contains N space-separated integers describing the respective elements of the array.

https://hackerrank-challenge-pdfs.s3.amazonaws.com/21237-s10-standard-deviation-English?AWSAccessKeyId=AKIAJ4WZFDFQTZRGO3QA&Expires=1554180474&Signature=9ZPaPXXHAPEIMHWPwrZKm4jVLdU%3D&response-content-disposition=inline%3B%20filename%3Ds10-standard-deviation-English.pdf&response-content-type=application%2Fpdf